Updated jump mechanics

+++ Megaman CMSC451/Megaman CMSC451/BackGround.cpp
@@ -63,7 +63,7 @@ void BackGround::drawView(GLint cmX)
 	Box b(cmX, Y, W, H, bg);
 	b.draw();
 }
-Rectangle2D * BackGround::getBelow(GLint x, GLint y)
+Rectangle2D * BackGround::getBelow(float x, float y)
 {
 	Box * ret = nullptr;
 
@@ -77,7 +77,7 @@ Rectangle2D * BackGround::getBelow(GLint x, GLint y)
 	return ret;
 }
 
-bool BackGround::canMove(GLint x, GLint y)
+bool BackGround::canMove(float x, float y)
 {
 	if(x < 0 || x > width) return false; // out of boundary
